+++ tests/search.spec.ts
@@ -1,0 +1,2 @@
+import { test, expect } from "@playwright/test";
+import { getApis } from "./helper";
