+++ tests/testsData/apisToTest/live-sensor.json
@@ -1,0 +1,17 @@
+[
+  {
+    "env": "test",
+    "name": "live sensor",
+    "url": "https://test-apps-ninja01.konturlabs.com/active/api/features/live-sensor"
+  },
+  {
+    "env": "dev",
+    "name": "live sensor",
+    "url": "https://test-apps-ninja02.konturlabs.com/active/api/features/live-sensor"
+  },
+  {
+    "env": "prod",
+    "name": "live sensor",
+    "url": "https://atlas.kontur.io/active/api/features/live-sensor"
+  }
+]
