+++ tests/testsData/apisToTest/oam-apis.json
@@ -1,0 +1,29 @@
+[
+  {
+    "env": "prod",
+    "name": "oam tiles",
+    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/308/257.png",
+    "urlLargeImg": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/308/257@2x.png",
+    "expectedImgAddress": "tests/testsData/oamTiles/256px-imageWithTiles.png",
+    "expectedLargeImgAddress": "tests/testsData/oamTiles/512px-imageWithTiles.png"
+  },
+  {
+    "env": "prod",
+    "name": "oam no tiles",
+    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/307/254.png",
+    "urlLargeImg": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/307/254@2x.png",
+    "expectedImgAddress": "tests/testsData/oamTiles/256px-imageWithNoTiles.png",
+    "expectedLargeImgAddress": "tests/testsData/oamTiles/512px-imageWithNoTiles.png"
+  },
+  {
+    "env": "prod",
+    "name": "cluster with tiles",
+    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/clusters/10/608/509.mvt",
+    "expectedNumImages": "7"
+  },
+  {
+    "env": "prod",
+    "name": "cluster with no tiles",
+    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/clusters/10/609/509.mvt"
+  }
+]
