+++ tests/testsData/apisToTest/ups-assets.json
@@ -1,47 +1,5 @@
 [
   {
-    "env": "prod",
-    "name": "oam tiles",
-    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/308/257.png",
-    "urlLargeImg": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/308/257@2x.png",
-    "expectedImgAddress": "tests/testsData/oamTiles/256px-imageWithTiles.png",
-    "expectedLargeImgAddress": "tests/testsData/oamTiles/512px-imageWithTiles.png"
-  },
-  {
-    "env": "prod",
-    "name": "oam no tiles",
-    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/307/254.png",
-    "urlLargeImg": "https://prod-raster-tiler.k8s-01.konturlabs.com/oam/mosaic/9/307/254@2x.png",
-    "expectedImgAddress": "tests/testsData/oamTiles/256px-imageWithNoTiles.png",
-    "expectedLargeImgAddress": "tests/testsData/oamTiles/512px-imageWithNoTiles.png"
-  },
-  {
-    "env": "prod",
-    "name": "cluster with tiles",
-    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/clusters/10/608/509.mvt",
-    "expectedNumImages": "7"
-  },
-  {
-    "env": "prod",
-    "name": "cluster with no tiles",
-    "url": "https://prod-raster-tiler.k8s-01.konturlabs.com/clusters/10/609/509.mvt"
-  },
-  {
-    "env": "test",
-    "name": "live sensor",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/features/live-sensor"
-  },
-  {
-    "env": "dev",
-    "name": "live sensor",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/features/live-sensor"
-  },
-  {
-    "env": "prod",
-    "name": "live sensor",
-    "url": "https://atlas.kontur.io/active/api/features/live-sensor"
-  },
-  {
     "env": "test",
     "name": "atlas about page",
     "url": "https://test-apps-ninja01.konturlabs.com/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d/assets/about.md"
@@ -220,162 +178,5 @@
     "env": "prod",
     "name": "smart-city privacy page",
     "url": "https://maps.kontur.io/active/api/apps/634f23f5-f898-4098-a8bd-09eb7c1e1ae5/assets/privacy.md"
-  },
-  {
-    "env": "prod",
-    "name": "atlas layers",
-    "url": "https://atlas.kontur.io/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d/layers"
-  },
-  {
-    "env": "prod",
-    "name": "atlas global layers",
-    "url": "https://atlas.kontur.io/active/api/layers/search/global",
-    "appId": "9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "prod",
-    "name": "atlas main",
-    "url": "https://atlas.kontur.io/active/api/apps/configuration"
-  },
-  {
-    "env": "prod",
-    "name": "atlas config",
-    "url": "https://atlas.kontur.io/active/static/config/appconfig.json"
-  },
-  {
-    "env": "prod",
-    "name": "intercom 3-party",
-    "url": "https://api-iam.intercom.io/messenger/web/ping",
-    "appId": "e59cl64z"
-  },
-  {
-    "env": "prod",
-    "name": "basemap",
-    "url": "https://api.kontur.io/layers/tiles/basemap/style_ninja_en.json"
-  },
-  {
-    "env": "prod",
-    "name": "atlas event feed",
-    "url": "https://atlas.kontur.io/active/api/events/user_feeds"
-  },
-  {
-    "env": "prod",
-    "name": "current user",
-    "url": "https://atlas.kontur.io/active/api/users/current_user"
-  },
-  {
-    "env": "prod",
-    "name": "app info",
-    "url": "https://atlas.kontur.io/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "prod",
-    "name": "llmAnalytics",
-    "url": "https://atlas.kontur.io/active/api/llm_analytics"
-  },
-  {
-    "env": "test",
-    "name": "atlas layers",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d/layers"
-  },
-  {
-    "env": "test",
-    "name": "atlas global layers",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/layers/search/global",
-    "appId": "9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "test",
-    "name": "atlas main",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/apps/configuration?appId=9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "test",
-    "name": "atlas config",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/static/config/appconfig.json"
-  },
-  {
-    "env": "test",
-    "name": "intercom 3-party",
-    "url": "https://api-iam.intercom.io/messenger/web/ping",
-    "appId": "e59cl64z"
-  },
-  {
-    "env": "test",
-    "name": "basemap",
-    "url": "https://test-basemap-tileserver.k8s-01.konturlabs.com/layers/tiles/basemap/style_ninja.json"
-  },
-  {
-    "env": "test",
-    "name": "atlas event feed",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/events/user_feeds"
-  },
-  {
-    "env": "test",
-    "name": "current user",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/users/current_user"
-  },
-  {
-    "env": "test",
-    "name": "app info",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "test",
-    "name": "llmAnalytics",
-    "url": "https://test-apps-ninja01.konturlabs.com/active/api/llm_analytics"
-  },
-
-  {
-    "env": "dev",
-    "name": "atlas layers",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d/layers"
-  },
-  {
-    "env": "dev",
-    "name": "atlas global layers",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/layers/search/global",
-    "appId": "9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "dev",
-    "name": "atlas main",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/apps/configuration?appId=9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "dev",
-    "name": "atlas config",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/static/config/appconfig.json"
-  },
-  {
-    "env": "dev",
-    "name": "intercom 3-party",
-    "url": "https://api-iam.intercom.io/messenger/web/ping",
-    "appId": "e59cl64z"
-  },
-  {
-    "env": "dev",
-    "name": "basemap",
-    "url": "https://dev-basemap-tileserver.k8s-01.konturlabs.com/layers/tiles/basemap/style_ninja.json"
-  },
-  {
-    "env": "dev",
-    "name": "atlas event feed",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/events/user_feeds"
-  },
-  {
-    "env": "dev",
-    "name": "current user",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/users/current_user"
-  },
-  {
-    "env": "dev",
-    "name": "app info",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/apps/9043acf9-2cf3-48ac-9656-a5d7c4b7593d"
-  },
-  {
-    "env": "dev",
-    "name": "llmAnalytics",
-    "url": "https://test-apps-ninja02.konturlabs.com/active/api/llm_analytics"
   }
 ]
